+++ src/CorruptedXdt.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Druc\XdtParser;
+
+class CorruptedXdt extends \Exception
+{
+    protected $message = 'File is corrupted';
+}